refactor(tests): extract interception e2e tests into a module
Moves two long integration tests from the monolithic interception_e2e.rs file into a dedicated submodule, keeping the main test file focused on the core interception scenarios. The extracted tests cover frame binding resolution and retry error reporting, which are self-contained and benefit from being grouped together.

Auto-committed-on: dragonfly
